Remove extra test file
