+++ tests/unit/test_messages.py
@@ -105,6 +105,47 @@ class MessageTestCase(unittest.TestCase):
         expected = [8, 4, 8259, {}, "light.bug", ["banana"]]
         self.assertEqual(error_message.value, expected)
 
+    def test_subscribe_message(self):
+        subscribe_message = wamp.SubscribeMessage(request_id=1395, topic="lesson.1")
+        self.assertEqual(subscribe_message.code, Code.SUBSCRIBE)
+        self.assertEqual(subscribe_message.request_id, 1395)
+        self.assertEqual(subscribe_message.options, {})
+        self.assertEqual(subscribe_message.topic, "lesson.1")
+        expected = [32, 1395, {}, "lesson.1"]
+        self.assertEqual(subscribe_message.value, expected)
+
+    def test_subscribed_message(self):
+        subscribed_message = wamp.SubscribedMessage(request_id=4872, subscription_id=653)
+        self.assertEqual(subscribed_message.code, Code.SUBSCRIBED)
+        self.assertEqual(subscribed_message.request_id, 4872)
+        self.assertEqual(subscribed_message.subscription_id, 653)
+        expected = [33, 4872, 653]
+        self.assertEqual(subscribed_message.value, expected)
+
+    def test_publish_message(self):
+        publish_message = wamp.PublishMessage(request_id=514, topic="zazie")
+        self.assertEqual(publish_message.code, Code.PUBLISH)
+        self.assertEqual(publish_message.request_id, 514)
+        self.assertEqual(publish_message.topic, "zazie")
+        expected = [16, 514, {}, "zazie"]
+        self.assertEqual(publish_message.value, expected)
+
+    def test_published_message(self):
+        published_message = wamp.PublishedMessage(request_id=681, publication_id=5092)
+        self.assertEqual(published_message.code, Code.PUBLISHED)
+        self.assertEqual(published_message.request_id, 681)
+        self.assertEqual(published_message.publication_id, 5092)
+        expected = [17, 681, 5092]
+        self.assertEqual(published_message.value, expected)
+
+    def test_event_message(self):
+        event_message = wamp.EventMessage(subscription_id=74, publication_id=27)
+        self.assertEqual(event_message.code, Code.EVENT)
+        self.assertEqual(event_message.subscription_id, 74)
+        self.assertEqual(event_message.publication_id, 27)
+        expected = [36, 74, 27, {}]
+        self.assertEqual(event_message.value, expected)
+
     def test_result_message(self):
         result_message = wamp.ResultMessage(request_id=8259)
         self.assertEqual(result_message.code, Code.RESULT)
